Include timezone and GMT offset in announcements
Since the meeting's been online we're often joined by folks in non-ET timezones.
To make scheduling easier for them, we ought to include the event's timezone and
GMT offset in the announcement.

This adds the timezone and GMT offset to event announcements, like so:

> Monday, May 3, 2021
> 7:00 PM EDT (GMT-0400)

It explicitly sets the site's timezone to "America/New_York" to ensure that,
even if the site is built on machines with different timezones, it'll treat
timestamps as being in ET.

Manual testing suggests that daylight savings time works the way we'd expect.